requirements.txt
deepseek-ai>=0.1.0
newsapi-python>=0.2.7
python-dotenv>=1.0.0
schedule>=1.2.1
jinja2>=3.1.3
requests>=2.31.0
python-dateutil>=2.8.2 